Standardize internal connection API
